requirements.txt
absl-py==1.0.0
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cmdstanpy==0.9.68
convertdate==2.3.2
cycler==0.11.0
Cython==0.29.26
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
ephem==4.1.3
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hijri-converter==2.2.2
holidays==0.11.3.1
idna==3.3
importlib-metadata==4.10.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
libclang==12.0.0
LunarCalendar==0.0.9
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multitasking==0.0.10
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
prophet==1.0.1
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
PyMeeus==0.5.11
Pympler==1.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==1.11.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
Send2Trash==1.8.0
setuptools-git==1.2
six==1.16.0
smmap==5.0.0
streamlit==1.3.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
urllib3==1.26.7
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yfinance==0.1.67
zipp==3.6.0
